// src/templates/back/packages/o-auth/src/@app/o-auth/client/application/upsert/o-auth-upsert-client.command.ts
import { CQMetadata } from '@aurorajs.dev/core';

export class OAuthUpsertClientCommand
{
    constructor(
        public readonly payload: {
            id: string;
            grantType?: string;
            name?: string;
            secret?: string;
            authUrl?: string;
            redirect?: string;
            scopeOptions?: any;
            expiredAccessToken?: number;
            expiredRefreshToken?: number;
            isActive?: boolean;
            isMaster?: boolean;
            applicationIds?: string[];
        },
        public readonly cQMetadata?: CQMetadata,
    ) {}
}